backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
yookassa==3.10.0
psutil==5.9.8
python-dateutil==2.8.2